Create index.php
Adding the index.php.  This will be the main page.  I need to add functions.php which will have php functions that get called by several php files (i.e., checking if a device is mobile or not).  Because this only has a PC template nothing will be displayed if someone looks at atomicprairie on a mobile site